+++ lib/flutter_in_app_pip.dart
@@ -4,3 +4,4 @@ export 'pip_capable_widget.dart';
 export 'picture_in_picture.dart';
 export 'pip_material_app.dart';
 export 'pip_widget.dart';
+export 'navigatable_pip_widget.dart';