+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.0
+* Removed unnecessary Navigator that was behaving as root navigator.
+* Added NavigatablePiPWidget which allows navigation inside the PiP Window.
+
 ## 1.1.0
 * Added ability to use builder in PiPMaterialApp.
 * Fixed a major bug where clicks near the boundary were not getting recognized in pip.